Move poco functions to namespace
Moved poco functions to Friendica\Protocol namespace. Update references and calls.
