+++ include/discover_poco.php
@@ -8,6 +8,7 @@ use Friendica\Core\Worker;
 use Friendica\Database\DBM;
 use Friendica\Model\GContact;
 use Friendica\Network\Probe;
+use Friendica\Protocol\PortableContact;
 
 require_once 'include/datetime.php';
 
@@ -21,7 +22,7 @@ function discover_poco_run(&$argv, &$argc)
 	- server <poco url>: Searches for the poco server list. "poco url" is base64 encoded.
 	- update_server: Frequently check the first 250 servers for vitality.
 	- update_server_directory: Discover the given server id for their contacts
-	- GContact::load: Load POCO data from a given POCO address
+	- PortableContact::load: Load POCO data from a given POCO address
 	- check_profile: Update remote profile data
 	*/
 
@@ -53,7 +54,7 @@ function discover_poco_run(&$argv, &$argc)
 
 	if ($mode == 8) {
 		if ($argv[2] != "") {
-			GContact::lastUpdated($argv[2], true);
+			PortableContact::lastUpdated($argv[2], true);
 		}
 	} elseif ($mode == 7) {
 		if ($argc == 6) {
@@ -61,9 +62,9 @@ function discover_poco_run(&$argv, &$argc)
 		} else {
 			$url = '';
 		}
-		GContact::loadWorker(intval($argv[2]), intval($argv[3]), intval($argv[4]), $url);
+		PortableContact::loadWorker(intval($argv[2]), intval($argv[3]), intval($argv[4]), $url);
 	} elseif ($mode == 6) {
-		GContact::discoverSingleServer(intval($argv[2]));
+		PortableContact::discoverSingleServer(intval($argv[2]));
 	} elseif ($mode == 5) {
 		update_server();
 	} elseif ($mode == 4) {
@@ -76,7 +77,7 @@ function discover_poco_run(&$argv, &$argc)
 			return;
 		}
 		$result = "Checking server ".$server_url." - ";
-		$ret = GContact::checkServer($server_url);
+		$ret = PortableContact::checkServer($server_url);
 		if ($ret) {
 			$result .= "success";
 		} else {
@@ -92,7 +93,7 @@ function discover_poco_run(&$argv, &$argc)
 		gs_search_user($search);
 	} elseif (($mode == 0) && ($search == "") && (Config::get('system', 'poco_discovery') > 0)) {
 		// Query Friendica and Hubzilla servers for their users
-		GContact::discover();
+		PortableContact::discover();
 
 		// Query GNU Social servers for their users ("statistics" addon has to be enabled on the GS server)
 		if (!Config::get('system', 'ostatus_disabled')) {
@@ -120,7 +121,7 @@ function update_server()
 	$updated = 0;
 
 	foreach ($r as $server) {
-		if (!GContact::doUpdate($server["created"], "", $server["last_failure"], $server["last_contact"])) {
+		if (!PortableContact::doUpdate($server["created"], "", $server["last_failure"], $server["last_contact"])) {
 			continue;
 		}
 		logger('Update server status for server '.$server["url"], LOGGER_DEBUG);
@@ -178,7 +179,7 @@ function discover_users()
 			continue;
 		}
 
-		$server_url = Gcontact::detectServer($user["url"]);
+		$server_url = PortableContact::detectServer($user["url"]);
 		$force_update = false;
 
 		if ($user["server_url"] != "") {
@@ -187,7 +188,7 @@ function discover_users()
 			$server_url = $user["server_url"];
 		}
 
-		if ((($server_url == "") && ($user["network"] == NETWORK_FEED)) || $force_update || GContact::checkServer($server_url, $user["network"])) {
+		if ((($server_url == "") && ($user["network"] == NETWORK_FEED)) || $force_update || PortableContact::checkServer($server_url, $user["network"])) {
 			logger('Check profile '.$user["url"]);
 			Worker::add(PRIORITY_LOW, "discover_poco", "check_profile", $user["url"]);
 
@@ -233,13 +234,13 @@ function discover_directory($search) {
 					continue;
 				}
 				// Update the contact
-				GContact::lastUpdated($jj->url);
+				PortableContact::lastUpdated($jj->url);
 				continue;
 			}
 
-			$server_url = GContact::detectServer($jj->url);
+			$server_url = PortableContact::detectServer($jj->url);
 			if ($server_url != '') {
-				if (!GContact::checkServer($server_url)) {
+				if (!PortableContact::checkServer($server_url)) {
 					logger("Friendica server ".$server_url." doesn't answer.", LOGGER_DEBUG);
 					continue;
 				}
