+++ mod/poco.php
@@ -9,7 +9,7 @@ use Friendica\App;
 use Friendica\Core\Cache;
 use Friendica\Core\Config;
 use Friendica\Database\DBM;
-use Friendica\Model\GContact;
+use Friendica\Protocol\PortableContact;
 
 function poco_init(App $a)
 {
@@ -37,7 +37,7 @@ function poco_init(App $a)
 
 	if ($a->argc > 1 && $a->argv[1] === '@server') {
 		// List of all servers that this server knows
-		$ret = GContact::serverlist();
+		$ret = PortableContact::serverlist();
 		header('Content-type: application/json');
 		echo json_encode($ret);
 		killme();
