+++ src/Worker/OnePoll.php
@@ -7,7 +7,7 @@ namespace Friendica\Worker;
 use Friendica\Core\Config;
 use Friendica\Core\PConfig;
 use Friendica\Database\DBM;
-use Friendica\Model\GContact;
+use Friendica\Protocol\PortableContact;
 use dba;
 
 require_once 'include/follow.php';
@@ -78,7 +78,7 @@ class OnePoll
 			);
 			if (DBM::is_result($r)) {
 				if (!$r[0]['total']) {
-					GContact::load($contact['id'], $importer_uid, 0, $contact['poco']);
+					PortableContact::load($contact['id'], $importer_uid, 0, $contact['poco']);
 				}
 			}
 		}
@@ -86,8 +86,8 @@ class OnePoll
 		/// @TODO Check why we don't poll the Diaspora feed at the moment (some guid problem in the items?)
 		/// @TODO Check whether this is possible with Redmatrix
 		if ($contact["network"] == NETWORK_DIASPORA) {
-			if (GContact::doUpdate($contact["created"], $contact["last-item"], $contact["failure_update"], $contact["success_update"])) {
-				$last_updated = GContact::lastUpdated($contact["url"]);
+			if (PortableContact::doUpdate($contact["created"], $contact["last-item"], $contact["failure_update"], $contact["success_update"])) {
+				$last_updated = PortableContact::lastUpdated($contact["url"]);
 				$updated = datetime_convert();
 				if ($last_updated) {
 					$fields = array('last-item' => $last_updated, 'last-update' => $updated, 'success_update' => $updated);
@@ -122,7 +122,7 @@ class OnePoll
 
 		// Update the contact entry
 		if (($contact['network'] === NETWORK_OSTATUS) || ($contact['network'] === NETWORK_DIASPORA) || ($contact['network'] === NETWORK_DFRN)) {
-			if (!GContact::reachable($contact['url'])) {
+			if (!PortableContact::reachable($contact['url'])) {
 				logger("Skipping probably dead contact ".$contact['url']);
 				return;
 			}
